requirements.txt
Pillow
matplotlib
RPI.GPIO
pigpio
adafruit-circuitpython-max31856
simple-pid
board
tk
gpiozero
